Add resolve_option function and refactor

+++ fv3config/_datastore.py
@@ -3,6 +3,7 @@ import tarfile
 import shutil
 import logging
 import appdirs
+from subprocess import check_call
 from ._exceptions import ConfigError, DataMissingError
 try:
     import wget
@@ -20,8 +21,9 @@ filename = '2019-10-23-data-for-running-fv3gfs.tar.gz'
 filename_root = '2019-10-23-data-for-running-fv3gfs'
 url = f'http://storage.googleapis.com/vcm-ml-public/{filename}'
 local_archive_filename = os.path.join(app_data_dir, filename)
+gs_bucket_prefix = 'gs://'
 
-forcing_directory_dict = {
+forcing_options_dict = {
     'default': os.path.join(local_archive_dir, 'base_forcing')
 }
 
@@ -67,17 +69,7 @@ def get_base_forcing_directory(config):
     """
     if 'forcing' not in config:
         raise ConfigError('config dictionary must have a \'forcing\' key')
-    forcing_option = config['forcing']
-    if os.path.isdir(forcing_option):
-        return_value = forcing_option
-    else:
-        if forcing_option not in forcing_directory_dict.keys():
-            raise ConfigError(
-                f'Forcing option {forcing_option} is not one of the valid options: {list(forcing_directory_dict.keys())}'
-            )
-        else:
-            return_value = forcing_directory_dict[forcing_option]
-    return return_value
+    return resolve_option(config['forcing'], forcing_options_dict)
 
 
 def get_initial_conditions_directory(config):
@@ -85,41 +77,34 @@ def get_initial_conditions_directory(config):
     """
     if 'initial_conditions' not in config:
         raise ConfigError('config dictionary must have an \'initial_conditions\' key')
-    if config['initial_conditions'] in initial_conditions_options_dict:
-        resolution = get_resolution(config)
-        if resolution != 'C48':
-            raise NotImplementedError(
-                'Default initial conditions only available for C48, please specify an initial conditions directory'
-            )
-        dirname = initial_conditions_options_dict[config['initial_conditions']]
-    elif os.path.isdir(config['initial_conditions']):
-        dirname = config['initial_conditions']
-    else:
-        raise ConfigError(
-            f'Initial conditions {config["initial_conditions"]} '
-            f'is not one of the valid options: {list(initial_conditions_options_dict.keys())}'
-        )
-    return dirname
+    return resolve_option(config['initial_conditions'], initial_conditions_options_dict)
 
 
 def link_directory(source_path, target_path):
-    """Recursively symbolic link the files in a source path into a target path.
+    """Recursively symbolic link the files in a source path into a target path or copy
+    from Google storage bucket to target path if source_path starts with gs://
     """
-    for base_filename in os.listdir(source_path):
-        source_item = os.path.join(source_path, base_filename)
-        target_item = os.path.join(target_path, base_filename)
-        if os.path.isfile(source_item):
-            if os.path.exists(target_item):
-                os.remove(target_item)
-            os.symlink(source_item, target_item)
-        elif os.path.isdir(source_item):
-            if not os.path.isdir(target_item):
-                os.mkdir(target_item)
-            link_directory(source_item, target_item)
+    if source_path.startswith(gs_bucket_prefix):
+        check_call(['gsutil', 'cp' '-r', source_path, target_path])
+    else:
+        for base_filename in os.listdir(source_path):
+            source_item = os.path.join(source_path, base_filename)
+            target_item = os.path.join(target_path, base_filename)
+            if os.path.isfile(source_item):
+                if os.path.exists(target_item):
+                    os.remove(target_item)
+                os.symlink(source_item, target_item)
+            elif os.path.isdir(source_item):
+                if not os.path.isdir(target_item):
+                    os.mkdir(target_item)
+                link_directory(source_item, target_item)
 
 
 def copy_file(source_path, target_path):
-    shutil.copy(source_path, target_path)
+    if source_path.startswith(gs_bucket_prefix):
+        check_call(['gsutil', 'cp', source_path, target_path])
+    else:
+        shutil.copy(source_path, target_path)
 
 
 def check_if_data_is_downloaded():
@@ -164,6 +149,40 @@ def extract_data():
     with tarfile.open(os.path.join(app_data_dir, filename), mode='r:gz') as f:
         f.extractall(app_data_dir)
         shutil.move(os.path.join(app_data_dir, filename_root), local_archive_dir)
+
+
+def resolve_option(option, built_in_options_dict):
+    """Determine whether an option is a built-in option or not and return path
+    to file or directory representing option. An option is assumed to be built-in
+    if it is not an absolute path and does not begin with gs://
+
+    Args:
+        option (str): an option
+        built_in_options_dict (dict): built-in options
+
+    Returns:
+        (str): a path or url
+
+    Raises:
+        ConfigError: if option is an absolute path but does not exist or if
+                     option is not in default_options_dict
+    """
+    if os.path.isabs(option):
+        if os.path.exists(option):
+            return option
+        else:
+            raise ConfigError(
+                f'The provided path {option} does not exist.'
+            )
+    elif option.startswith(gs_bucket_prefix):
+        return option
+    else:
+        if option in built_in_options_dict:
+            return built_in_options_dict[option]
+        else:
+            raise ConfigError(
+                f'The option {option} is not one of the built in options: {list(built_in_options_dict.keys())}'
+            )
 
 
 if __name__ == '__main__':
